add dbrepo for register,login,article publish, article saveas, article edit, article delete, article getall, article getone, article search, comment insert, comment getall, comment delete, etc....

+++ mock/mock-articles.go
@@ -10,7 +10,7 @@ func returnRandomArticleID() models.Article {
 	randID := rand.Int31()
 
 	return models.Article{
-		ID:       uint(randID),
+		ID:       int64(randID),
 		Date:     "17th May 2021",
 		Title:    "What is Cryptocurrency",
 		Preview:  "Cryptocurrency is decentralized digital money, based on blockchain technology. You may be familiar with the most popular versions, Bitcoin and Ethereum, but there are more than 5,000 different cryptocurrencies in circulation, according to CoinLore. You can use crypto to buy regular goods and services, although many people invest in cryptocurrencies as they would in other assets, like stocks or precious metals. While cryptocurrency is a novel and exciting asset class, purchasing it can be risky as you must take on a fair amount of research to fully understand how each system works.",
