+++ models/auth.go
@@ -2,7 +2,7 @@ package models
 
 //AuthFromServer is how auth from client structured
 type AuthFromClient struct {
-	ID    uint
+	ID    int64
 	Name  string
 	Pass  string
 	Token string
@@ -17,7 +17,7 @@ type AuthFromServer struct {
 
 //Settings is how settings format be sent or received
 type Settings struct {
-	ID        uint
+	ID        int64
 	Name      string
 	Bio       string
 	AvatarURL string
