+++ models/comment.go
@@ -2,14 +2,14 @@ package models
 
 //Comment is how comment structured
 type Comment struct {
-	ID   uint
+	ID   int64
 	Name string
 	Text string
 }
 
 //CommentFromClient is how comment from client structured
 type CommentFromClient struct {
-	ID                uint
+	ID                int64
 	CommentFromClient Comment
 }
